LayoutTests/imported/w3c/web-platform-tests/css/css-ui-3/outline-017-expected.txt


PASS outline-color is animated as a color 
PASS outline-width is animated as a length 
PASS outline-offset is animated as a length 
